Remove generic constraints on IMessageHandler interface
Requiring class, new() means that c# record types can't be used for the message
type, which sucks because records are very good for holding data which is what
messsages do.

In order to get the tests to build I had to add the IsExternalInit method -
which afaik shouldn't be required for .net5 so it's very confusing, but this
should be ok because it's just the test project so it won't be referenced by
anything else.